Task 4) Final: Update function valorMasGrandeDeLasPropiedadesDeUnObjetoConNumeros
With this last step, I can get the object which has the highest -numeric- value as property. For this reason, now I can make a better code to get 4) task assigment done.

Also, I deleted //Pruebas (testing, console.log statements) and all other variable that isn't affecting the result of the different tasks.

Besides, there is a function that is "repeated": nodeToArray. Actually, it's not repeated since there isn't any other function that does the same thing BUT there is a for loop at line 13 which does the same thing.
The reason why I decided to keep the function is that I had to apply the .forEach method, so I learnt it by building this function.